feat: reduce the public api for resource builder
